rename visual_paneldates
