+++ macrosynergy/visuals/timelines.py
@@ -302,7 +302,7 @@ def timelines(
 
 
 if __name__ == "__main__":
-    from macrosynergy.visuals.plots import FacetPlot
+    from macrosynergy.visuals import FacetPlot
     from macrosynergy.management.simulate_quantamental_data import make_test_df
 
     cids: List[str] = [
